requirements.txt
# 核心依賴
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 數據分析和視覺化
matplotlib>=3.7.0
seaborn>=0.12.0
talib>=0.4.25

# 機器學習和技術分析
numpy>=1.24.0

# LINE Bot SDK
line-bot-sdk>=3.5.0

# OpenAI API
openai>=1.0.0

# Jupyter 支持 (可選)
ipython>=8.14.0
jupyter>=1.0.0
notebook>=6.5.0

# 其他工具
python-dotenv>=1.0.0
certifi>=2023.0.0

# 開發工具 (可選)
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0